Move output model interface to mode.ts
